% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_offsets.R
\name{add_cor_offset}
\alias{add_cor_offset}
\title{Add 'correct offset' column to a set of reads}
\usage{
add_cor_offset(rpfs, anno)
}
\arguments{
\item{rpfs}{GRanges object with RPFs}

\item{anno}{annotation object}
}
\value{
the rpf granges but wth a 'cor_offset' column
}
\description{
Given rpfs, and named vectors of starts and stops, this adds the 'correct'
offset to the reads to use with determining p-site offsets
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
