% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ritpms.R
\name{addphase}
\alias{addphase}
\title{Add 'phase' column to a set of reads, given a vector of cds starts}
\usage{
addphase(gr, cdsstarts)
}
\arguments{
\item{cdsstarts}{named vector of cds starts}

\item{rpfs}{GRanges object with RPFs}
}
\value{
the rpf granges but wth a 'phase' column
}
\description{
Given rpfs, and named vectors of starts and stops, this adds the 'phase'
to the reads
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
