% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr22_anno}
\alias{chr22_anno}
\title{Gencode v32 annotation for human chromosome 22}
\format{
An annotation object as created by load_annotation:
}
\source{
\url{AnnotationHub gencode v32 annotation}
}
\usage{
chr22_anno
}
\description{
The gencode v32 annotation, downloaded from Annotationhub()
and then processed with load_annotation, including uORFs
}
\keyword{datasets}
