% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{convert_gtf}
\alias{convert_gtf}
\title{Convert old-style gtfs (e.g. those output by rtracklayer) to richer format}
\usage{
convert_gtf(anno, keep_cols)
}
\arguments{
\item{anno}{annotation GRanges read in by rtracklayer}

\item{keep_cols}{the columns to keep in the final annotation}
}
\value{
a granges object with just the exon and CDS elements
that ribostan uses
}
\description{
Convert old-style gtfs (e.g. those output by rtracklayer) to richer format
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
