% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_div.R
\name{export_codon_dts}
\alias{export_codon_dts}
\title{Create a dataframe with a_site, p_site and e site occupancies}
\usage{
export_codon_dts(metacodondf, kl_offsets)
}
\arguments{
\item{metacodondf}{A data frame with expected and actual rust score per position in
window, codon, readlength}

\item{kl_offsets}{A dataframe with p-site offsets per readlength, sample}
}
\value{
a dataframe occupancies for e,p and a sites for each sample
}
\description{
Create a dataframe with a_site, p_site and e site occupancies
}
\examples{
data(chr22_anno)
data(rpfs)
data(offsets_df)
data(metacodondf)
# note this doesn't work that well on a small subset
kl_df <- get_kl_df(metacodondf, chr22_anno)
kl_offsets <- select_offsets(kl_df)
allcodondt <- export_codon_dts(metacodondf, kl_offsets)
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
