% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{fmcols}
\alias{fmcols}
\title{Pick columns from a GRangestList}
\usage{
fmcols(grl, ...)
}
\arguments{
\item{grl}{String; full path to html report file.}

\item{...}{metadata column to pull}
}
\value{
length n vector pulled from mcols of first list elements
}
\description{
Given a grangelist of say N genes with X_n exons, this yields a
length N vector pulled from the mcols of the first element of each list
element
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
