% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity_tests.R
\name{ftest_orfs}
\alias{ftest_orfs}
\title{Run Multitaper Tests on All ORFs}
\usage{
ftest_orfs(psites, anno, n_cores = 1)
}
\arguments{
\item{psites}{A GRanges object containing psites}

\item{anno}{annotation object}

\item{n_cores}{number of cores to use}
}
\value{
a numeric vector with the spectral coefficient at 0.333...
and the pvalue for the test
}
\description{
Run Multitaper Tests on All ORFs
}
\details{
This function applies a multitaper test to
}
\examples{
data(chr22_anno)
data(rpfs)
data(offsets_df)
psites <- get_psite_gr(rpfs, offsets_df, chr22_anno)
ftests <- ftest_orfs(psites \%>\% head(10000), chr22_anno, 
  n_cores=1)
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
