% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity_tests.R
\name{ftestvect}
\alias{ftestvect}
\title{Test a numeric vector for periodicity using the multitaper package}
\usage{
ftestvect(psit, k = 24, bw = 12)
}
\arguments{
\item{psit}{A list of SimpleRleLists - RPF coverage split by readlength}

\item{k}{\itemize{
\item The number of slepians to use for the test
}}

\item{bw}{\itemize{
\item The bandwidth to use for the test
}}
}
\value{
a numeric vector with the spectral coefficient at 0.333... and the pvalue for the test
}
\description{
Test a numeric vector for periodicity using the multitaper package
}
\details{
This function pads short vectors up to 50, and performs a multitaper test for 3bp
peridicity on it's input, returning a spectral coefficient and p-value
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
