% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_TE.R
\name{gene_level_elong}
\alias{gene_level_elong}
\title{Aggregate transcript elongation scores to gene level}
\usage{
gene_level_elong(tr_elong, ritpms, anno, exclude_uORFs = TRUE)
}
\arguments{
\item{tr_elong}{elongation rates per transcript, data frame with columns
orf_id, mean_occ}

\item{ritpms}{a named vector of ritpms per transcript}

\item{anno}{an annotation object}

\item{exclude_uORFs}{exclude uORFs from the gene level total;Defaults to TRUE}
}
\value{
a data frame with columns mean_occ and gene_id
}
\description{
Aggregate transcript elongation scores to gene level
}
\details{
This function aggregates transript level ritpms to gene level,with
the mean elongation rate over all the
genes transcripts weighted by transcript abundance
}
\seealso{
\code{\link{get_cds_reads}}, \code{\link{get_readlens}}
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
