% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ritpms.R
\name{gene_level_expr}
\alias{gene_level_expr}
\title{Aggregate transcript level RiboPMs to the gene-level
Ignores uORF expression}
\usage{
gene_level_expr(ripms, anno)
}
\arguments{
\item{ripms}{\itemize{
\item named vector of ribosome densities
}}

\item{anno}{An annotation object with a gene-transcript table}
}
\value{
a data frame with columns gene_id, expr
}
\description{
Aggregate transcript level RiboPMs to the gene-level
Ignores uORF expression
}
\details{

}
\examples{
data(chr22_anno)
data(rpfs)
data(offsets_df)
data(ms_df)

psites <- get_psite_gr(rpfs, offsets_df, chr22_anno)
# now, use Stan to estimate normalized p-site densities for our data
ritpms <- get_ritpms(psites, chr22_anno)
# and get these at the gene level (ignoring uORFs)
gritpms <- gene_level_expr(ritpms, chr22_anno)
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
