% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_div.R
\name{get_cds_codons}
\alias{get_cds_codons}
\title{Get A Granges object with locations of each codon in the given cds}
\usage{
get_cds_codons(
  anno,
  n_wind_l_ext = 45,
  n_wind_r_ext = 9,
  n_start_buff = 60,
  n_end_buff = 60
)
}
\arguments{
\item{anno}{An annotation object}

\item{n_wind_l_ext}{The 5' extent of the window around the codon to return
Defaults to: 45}

\item{n_wind_r_ext}{The 3' extent of the window around the codon to return
Defaults to: 9}

\item{n_start_buff}{The number of bp at the start of each cds to exclude
Defaults to 60}

\item{n_end_buff}{The number of bp at the end of each cds to exclude
Defaults to 60}
}
\value{
a GRanges object with the codon identity in the names, and the
}
\description{
Get A Granges object with locations of each codon in the given cds
}
\details{
Use estimates of per codon dwell time to get the mean dwell time
over transcrips
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
