% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{get_cdsgrl}
\alias{get_cdsgrl}
\title{Get a set of filtered cds from an imported GTF GRanges}
\usage{
get_cdsgrl(filt_anno, fafileob, ignore_orf_validity)
}
\arguments{
\item{filt_anno}{GRanges; an unfilted imported GTF}

\item{fafileob}{FaFile object; reference to an indexed genomic
fasta file}

\item{ignore_orf_validity}{Boolean; flag whether to include ORFs
missing a valid stop codon}
}
\value{
a GRangesList split by transcript, which contains the filtered coding
ranges for each one
}
\description{
Get a set of filtered cds from an imported GTF GRanges
}
\details{
This takes only coding sequences which are a multiple of 3bp and
have a start and a stop on either end.it always returns coding sequences
without the stops, regardless of their extent in the input.
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
