% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_TE.R
\name{get_codon_occs}
\alias{get_codon_occs}
\title{Estimate per codon occupancies}
\usage{
get_codon_occs(
  psites,
  offsets_df,
  anno,
  n_genes = 1000,
  covthresh = 10,
  method = "linear"
)
}
\arguments{
\item{psites}{\itemize{
\item a data frame of inferred RPF p-sites
}}

\item{offsets_df}{\itemize{
\item a data frame with numeric columns readlen,phase,offset
}}

\item{anno}{\itemize{
\item an annotation object
}}

\item{n_genes}{\itemize{
\item how many genes to use for the model - most highly expressed
used first
}}

\item{covthresh}{\itemize{
\item how many reads to require in an ORF for it to be used
defaults to RUST
}}

\item{method}{\itemize{
\item which method to use to estimate the codon occupancies,
}}
}
\value{
a data frame with columns codon,position,estimate,upper,lower,
p.value
}
\description{
Uses sampled RPF coverage gr object, offsets data frame, and an annotation
object
to get estimates of per-codon occupancy (which are proportional to
dwell times)
}
\examples{
data(chr22_anno)
data(rpfs)
data(offsets_df)
psites <- get_psite_gr(rpfs, offsets_df, chr22_anno)
rust_codon_occ_df <- get_codon_occs(psites, offsets_df, chr22_anno,
  n_genes = 200, method = "RUST"
)
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
