% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_TE.R
\name{get_codposdf}
\alias{get_codposdf}
\title{Generate a data frame of orf_id,position,codon}
\usage{
get_codposdf(orfnames, anno)
}
\arguments{
\item{orfnames}{\itemize{
\item the ORFs to get sequence data for
}}

\item{anno}{\itemize{
\item an annotation object
}}
}
\value{
returns a data frame with columns orf_id, pos, codon
}
\description{
Generate a data frame of orf_id,position,codon
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
