% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_div.R
\name{get_cov_rust_scores}
\alias{get_cov_rust_scores}
\title{Get A Granges object with locations of each codon in the given cds}
\usage{
get_cov_rust_scores(rpf_cov, cds_codons)
}
\arguments{
\item{rpf_cov}{A list of SimpleRleLists - RPF coverage split by readlength}

\item{cds_codons}{\itemize{
\item GRanges object containing windows in which to get rust scores
}}
}
\value{
A data frame with expected and actual rust score per position in window, codon, readlength
}
\description{
Get A Granges object with locations of each codon in the given cds
}
\details{
Use estimates of per codon dwell time to get the mean dwell time over transcrips
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
