% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ritpms.R
\name{get_exprfile}
\alias{get_exprfile}
\title{Given a bam file, and transcript fasta, output a file of ribosome densities}
\usage{
get_exprfile(ribobam, ribofasta, outfile)
}
\arguments{
\item{ribobam}{A bam file with RPFs}

\item{ribofasta}{A gencode style fasta to which RPFs were aligned}

\item{outfile}{the file to which the RPF densitites should be saved}
}
\value{
the file to which the table was saved
}
\description{
Given a bam file, and transcript fasta, output a file of ribosome densities
}
\details{
The Ribosome densities are saved in salmon format
}
\examples{
\dontrun{iget_exprfile(ribobam, ribofasta, outfile)}
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
