% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_offsets.R
\name{get_incl_max_offsets}
\alias{get_incl_max_offsets}
\title{Get Offsets by maximizing p-sites within the CDS}
\usage{
get_incl_max_offsets(rpfs, anno)
}
\arguments{
\item{rpfs}{GRanges object with RPFs}

\item{anno}{annotation object}
}
\value{
a data frame with the best offset per read/phase
}
\description{
Get Offsets by maximizing p-sites within the CDS
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
