% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_offsets.R
\name{get_offsets}
\alias{get_offsets}
\title{Get Offsets by maximizing p-sites within the CDS}
\usage{
get_offsets(input, anno, method = "cds_incl")
}
\arguments{
\item{input}{GRanges object with RPFs or a bam file with RPFs}

\item{anno}{annotation object}

\item{method}{Method for p-site determination. Defaults to 'cds_incl'}
}
\value{
a dataframe with p-site offsets per readlength
}
\description{
Get Offsets by maximizing p-sites within the CDS
}
\examples{
cov <- get_readgr(ribobam, anno)
offsets_df <- get_offsets(cov, anno)
data(chr22_anno)
data(rpfs)
data(offsets_df)
testoffsets_df <- get_offsets(rpfs, chr22_anno)

}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
