% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_TE.R
\name{get_orf_elong}
\alias{get_orf_elong}
\title{Get the mean dwell time over all translated codons for a set of transcripts}
\usage{
get_orf_elong(anno, codon_model)
}
\arguments{
\item{anno}{\itemize{
\item an annotation object
}}

\item{codon_model}{\itemize{
\item a dwell-time data frame with columns position, codon,
estimate
}}
}
\value{
a data frame with columns orf_id, mean_occ
}
\description{
Get the mean dwell time over all translated codons for a set of transcripts
}
\details{
Use estimates of per codon dwell time to get the mean dwell time
}
\examples{


data(chr22_anno)
data(rpfs)
data(offsets_df)
psites <- get_psite_gr(rpfs, offsets_df, chr22_anno)
rust_codon_occ_df <- get_codon_occs(psites, offsets_df, chr22_anno,
  n_genes = 1000, method = "RUST"
)
orf_elong <- get_orf_elong(chr22_anno, rust_codon_occ_df)
gn_elong <- gene_level_elong(orf_elong, ritpms, anno)
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
