% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ritpms.R
\name{get_psite_gr}
\alias{get_psite_gr}
\title{convert a granges object of footprints to a psites object}
\usage{
get_psite_gr(rpfs, offsets_df, anno)
}
\arguments{
\item{rpfs}{\itemize{
\item granges object with positions of footprints, in transcript
space
}}

\item{offsets_df}{\itemize{
\item a data frame with numeric columns readlen,phase,offset
}}

\item{anno}{\itemize{
\item an annotation object with at least one ORF per transcript,
the function uses the 'uORF' slot if it's available
}}
}
\value{
a GRanges object of width-1 psites with columns readlen,
phase and offset
}
\description{
This function applies offsets to RPF data, attributing to each RPF
an offset specific a readlength and phase. Where phase is ambigous,
because more than one ORF overlaps the RPF, the RPF has both offsets
applied, and in rare cases where more than one possible psite location
applies, one is randomly selected.
}
\examples{
data(chr22_anno)
data(rpfs)
data(offsets_df)
psites <- get_psite_gr(rpfs, offsets_df, anno)
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
