% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_TE.R
\name{get_psitecov}
\alias{get_psitecov}
\title{Get a coverage RleList of psite coverage over a set of ORFs}
\usage{
get_psitecov(rpfs, offsets_df, anno)
}
\arguments{
\item{rpfs}{\itemize{
\item granges object with positions of footprints, in transcript
space
}}

\item{offsets_df}{\itemize{
\item a data frame with numeric columns readlen,phase,offset
}}

\item{anno}{\itemize{
\item an annotation object with at least one ORF per transcript
}}
}
\value{
a numeric SimpleRleList for each ORF with p-site density
}
\description{
Get a coverage RleList of psite coverage over a set of ORFs
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
