% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ritpms.R
\name{get_read_spmat}
\alias{get_read_spmat}
\title{get_read_spmat}
\usage{
get_read_spmat(psites, anno)
}
\arguments{
\item{psites}{GRanges object of}

\item{anno}{an annotation object}
}
\value{
A matrix of the infile
}
\description{
get optimal ritpms using stan
}
