% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ritpms.R
\name{get_readgr}
\alias{get_readgr}
\title{Read and filter a bam file of RPFs}
\usage{
get_readgr(
  ribobam,
  anno,
  offsets_df = NULL,
  startstop = FALSE,
  strip_seqnames = TRUE
)
}
\arguments{
\item{ribobam}{GRanges; A bam file with RPFs}

\item{anno}{An annotation object}

\item{offsets_df}{\itemize{
\item a data frame with numeric columns readlen,phase,offset
}}

\item{startstop}{option to select only reads that overlapt he start/stop
for offset determination,Defaults to FALSE}

\item{strip_seqnames}{whether the function should remove all text after the
first '|', useful if aligning to gencode fastas Defaults to \code{TRUE}}
}
\value{
This function returns a granges object with the read names in the
names slot and a metadata column
denoting readlength.
}
\description{
Read and filter a bam file of RPFs
}
\details{
This function reads in a bam file containing ribosomal footprints.
It uses only reads without splice sites and reads which align to the
positive strand, as it's designed to work on transcriptomic alignments.
}
\examples{
data(chr22_anno)
testbam <- system.file("extdata", "chr22.bam", package = "Ribostan", mustWork = TRUE)
rpfs <- get_readgr(testbam, chr22_anno)
}
\seealso{
\code{\link{get_cds_reads}}, \code{\link{get_readlens}}
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
