% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ritpms.R
\name{get_readlens}
\alias{get_readlens}
\title{Select those read lengths which make up 95\% of the reads}
\usage{
get_readlens(readlens)
}
\arguments{
\item{readlens}{Numeric; numeric vector}
}
\value{
a numeric integer vector, such that selecting reads with these
lengths will preserve at least 95\%
of reads
}
\description{
This function takes in a vector of the lengths of each read, and determines
which read lengths should be included in order to filter out the 2.5\%
shortest/longest reads.
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
