% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{get_ribofasta_anno}
\alias{get_ribofasta_anno}
\title{This creates a minimal annotation object from a gencode style fasta}
\usage{
get_ribofasta_anno(ribofasta)
}
\arguments{
\item{ribofasta}{A gencode style fasta to which RPFs were aligned}
}
\value{
An annotation object with cdsstarts/stops
}
\description{
This creates a minimal annotation object from a gencode style fasta
}
\details{
The Ribosome densities are saved in salmon format
}
\examples{
gtf <- system.file("extdata", "gcv37.anno.chr22.gtf", package = "Ribostan", mustWork = TRUE)
fafile <- system.file("extdata", "chr22.fa.gz", package = "Ribostan", mustWork = TRUE)
file.copy(fafile, ".")
system2("gunzip -f chr22.fa.gz")
fafile <- "chr22.fa"
ext_fasta <- make_ext_fasta(gtf, fafile, outfasta = "tmp.fa", fpext = 50, tpext = 50)
get_ribofasta_anno(ext_fasta)
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
