% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ritpms.R
\name{get_ritpms}
\alias{get_ritpms}
\title{get_ritpms}
\usage{
get_ritpms(psites, anno)
}
\arguments{
\item{psites}{A GRanges object containing RPFs}

\item{anno}{An annotation object with a gene-transcript table}
}
\value{
a vector of normalized footprint densities
}
\description{
This function takes in a sparse matrix, and then samples a single value
from each column, with sampling weights within each column equal to
the columns values
}
\examples{
data(chr22_anno)
data(rpfs)
data(offsets_df)
data(ms_df)

psites <- get_psite_gr(rpfs, offsets_df, chr22_anno)
# now, use Stan to estimate normalized p-site densities for our data
ritpms <- get_ritpms(psites, chr22_anno)
}
