% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_TE.R
\name{get_sitedf}
\alias{get_sitedf}
\title{create a data frame with a and p site codons, and rpf counts per
transcript/position}
\usage{
get_sitedf(psite_cov, anno)
}
\arguments{
\item{psite_cov}{\itemize{
\item a data frame of RPF footprints
}}

\item{anno}{\itemize{
\item an annotation object
}}
}
\value{
a data frame with columns orf_id, p_codon, a_codon, count
}
\description{
create a data frame with a and p site codons, and rpf counts per
transcript/position
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
