% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{get_trspace_cds}
\alias{get_trspace_cds}
\title{Get CDS positions in transcript space}
\usage{
get_trspace_cds(cdsgrl, exonsgrl)
}
\arguments{
\item{cdsgrl}{GRangesList; List of filtered CDS GRanges from GTF
annotation}

\item{exonsgrl}{GRangesList; exons making up the space element to be mapped
from}
}
\value{
a granges object containing the coding sequence range for each
transcript
}
\description{
Get CDS positions in transcript space
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
