% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{is_out_of_bounds}
\alias{is_out_of_bounds}
\title{Check if GRanges elements are out of chromosome bounds}
\usage{
is_out_of_bounds(gr, si = seqinfo(gr))
}
\arguments{
\item{gr}{GRanges or GrangesList of elements}

\item{si}{seqinfo information about length and names of
chromosomes}
}
\value{
a logical vector valued TRUE if GRanges
elements are out of the chromosome bounds
}
\description{
Given a grangelist of say N genes with X_n exons, this yields a
length N vector pulled from the mcols of the first element of each
list element
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
