% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{load_annotation}
\alias{load_annotation}
\title{Get a set of filtered cds from an imported GTF GRanges}
\usage{
load_annotation(
  gtf,
  fafile,
  add_uorfs = TRUE,
  ignore_orf_validity = FALSE,
  keep_cols = c("gene_id", "transcript_id", "gene_name", "type"),
  DEFAULT_CIRC_SEQS = NULL,
  findUORFs_args = c(minimumLength = 0)
)
}
\arguments{
\item{gtf}{A GTF with gene annotation}

\item{add_uorfs}{Whether to look for and include uORFs}

\item{keep_cols}{columns to save from the gtf metadata}

\item{DEFAULT_CIRC_SEQS}{default chromsoomes to treat as circular}

\item{findUORFs_args}{Additional arguments to pass to ORFik::findUORFs}

\item{fafileob}{FaFile; a genomic Fasta file}
}
\value{
a list containing annotation objects used by other functions.
}
\description{
Get a set of filtered cds from an imported GTF GRanges
}
\details{
This takes only coding sequences which are a multiple of 3bp and
have a start and a stop on either end. it always returns coding sequences
without the stops, regardless of their extent in the input.
}
\examples{
gtf <- system.file("extdata", "gcv37.anno.chr22.gtf",
  package = "Ribostan",
  mustWork = TRUE
)
fafile <- system.file("extdata", "chr22.fa.gz",
  package = "Ribostan",
  mustWork = TRUE
)
file.copy(fafile, ".")
system2("gunzip -f chr22.fa.gz")
fafile <- "chr22.fa"
anno <- load_annotation(gtf, fafile)
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
