% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{make_ext_fasta}
\alias{make_ext_fasta}
\title{Create a fasta file of coding sequences extended nbp on either end.}
\usage{
make_ext_fasta(gtf, fasta, outfasta, fpext = 50, tpext = 50)
}
\arguments{
\item{gtf}{A GTF file with annotation.}

\item{fasta}{A genomic Fasta file}

\item{outfasta}{The name of the fasta file to output}

\item{fpext}{How many bp to extendt the CDS on the 5' end}

\item{tpext}{How many bp to extendt the CDS on the 3' end}
}
\value{
the name of the file to which the sequences were output
}
\description{
Create a fasta file of coding sequences extended nbp on either end.
}
\details{
This creates a fasta file with gencode-style headers. The sequences
are created by extending the coding sequences by the specified amount in
transcript space, (or past the) end of the transcript if necessary, so that
all sequences have the same 'UTRs'.
}
\examples{
gtf <- system.file("extdata", "gcv37.anno.chr22.gtf",
  package = "Ribostan",
  mustWork = TRUE
)
fafile <- system.file("extdata", "chr22.fa.gz",
  package = "Ribostan",
  mustWork = TRUE
)
file.copy(fafile, ".")
system2("gunzip -f chr22.fa.gz")
fafile <- "chr22.fa"
ext_fasta <- make_ext_fasta(gtf, fafile, outfasta = "tmp.fa", fpext = 50, tpext = 50)
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
