% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metacodondf}
\alias{metacodondf}
\title{A data frame containing metacodon data}
\format{
A tibble: with 23,058 rows and 8 columns
\describe{
\item{sample}{the sample used to calculate the metacodons}
\item{codon}{the codon used to calculate the metacodon profile}
\item{position}{the position relative tot he codon}
\item{ro_cl}{The RUST score at that position}
\item{re_c}{The expected RUST score at that position}
\item{count}{the mean read 5' end count at that position}
}
}
\source{
\url{AnnotationHub gencode v32 annotation}
}
\usage{
metacodondf
}
\description{
A data frame containing metacodon data
}
\keyword{datasets}
