% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_div.R
\name{most_freq}
\alias{most_freq}
\title{Get the most frequent element from a numeric vector}
\usage{
most_freq(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the single most common value in the vector
}
\description{
Get the most frequent element from a numeric vector
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
