% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{offsets_df}
\alias{offsets_df}
\title{A Dataframe containing CDS-occupancy derived offsets
for the RPF data in rpfs}
\format{
a  tibble with 12 rows and 3 columns
\describe{
\item{readlen}{the readlength to which the offset shoudl be applied}
\item{phase}{the phase to which the offset shoudl be applied}
\item{offset}{the offset which should be applied}
}
}
\source{
\url{AnnotationHub gencode v32 annotation}
}
\usage{
offsets_df
}
\description{
A Dataframe containing CDS-occupancy derived offsets
for the RPF data in rpfs
}
\keyword{datasets}
