% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ritpms.R
\name{optimize_ritpms}
\alias{optimize_ritpms}
\title{Use a sparse mapping matrix to optimize ritpms}
\usage{
optimize_ritpms(spmat, anno, iternum = 500, verbose = FALSE)
}
\arguments{
\item{spmat}{a sparse numeric matrix}

\item{anno}{an annotation object}

\item{iternum}{how many iterations to optimize TPMs for;
Defaults to 500}

\item{verbose}{whether to show optimization messages from stan;
Defaults to FALSE}
}
\value{
A matrix of the infile
}
\description{
get optimal ritpms using stan
}
