% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{order_grl_st}
\alias{order_grl_st}
\title{Index vector for a GRanges list object with the sub-elements ordered 5' to 3'}
\usage{
order_grl_st(grl)
}
\arguments{
\item{grl}{GRangesList; a GRangesList object}
}
\value{
Index vector for a GRanges list object with the sub-elements ordered
5' to 3'
}
\description{
Index vector for a GRanges list object with the sub-elements ordered 5' to 3'
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
