% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_div.R
\name{plot_kl_dv}
\alias{plot_kl_dv}
\title{Plot KL divergence in selected windows}
\usage{
plot_kl_dv(kl_df, kl_offsets, selreadlens = 27:32)
}
\arguments{
\item{kl_df}{dataframe with KL divergence per readlength, position in window
around codon}

\item{kl_offsets}{dataframe with information on p-sites}

\item{selreadlens}{integer vector of readlengths to plot}
}
\value{
a dataframe with p-site offsets per readlength
}
\description{
Plot KL divergence in selected windows
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
