% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{resize_grl_endfix}
\alias{resize_grl_endfix}
\title{Resize a GRangesList object holding it's 3' end fixed}
\usage{
resize_grl_endfix(grl, width)
}
\arguments{
\item{grl}{GRangesList; a GRangesList object}
}
\value{
A GRangesList object shortend/lengthened, respecting exon boundaries
}
\description{
Resize a GRangesList object holding it's 3' end fixed
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
