% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{resize_grl_startfix}
\alias{resize_grl_startfix}
\title{Resize a GRangesList object holding it's 5' end fixed}
\usage{
resize_grl_startfix(grl, width)
}
\arguments{
\item{grl}{GRangesList; a GRangesList object}

\item{width}{GRangesList; integer/IntegerList to set as new width.}
}
\value{
A GRangesList object shortend/lengthened, respecting exon boundaries
}
\description{
Resize a GRangesList object holding it's 5' end fixed
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
