% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rpfs}
\alias{rpfs}
\title{GRanges object with 63346 ranges and 1 metadata column:}
\format{
GRanges object with 63346 ranges and 1 metadata column:
\describe{
\item{readlen}{the readlength of the footprint}
}
}
\source{
\url{Simulated data}
}
\usage{
rpfs
}
\description{
GRanges object with 63346 ranges and 1 metadata column:
}
\keyword{datasets}
