% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ritpms.R
\name{sample_cols_spmat}
\alias{sample_cols_spmat}
\title{sample_cols_spmat}
\usage{
sample_cols_spmat(spmat, return_mat = TRUE)
}
\arguments{
\item{spmat}{a sparse numeric matrix describing RPF's multimapping}

\item{return_mat}{whether to a matrix for selecting the elements,
as opposed to the summary data frame;defaults to TRUE}
}
\value{
A matrix of the
}
\description{
This function takes in a sparse matrix, and then samples a single value
from each column, with sampling weights within each column equal to
the columns values
}
