% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ritpms.R
\name{sample_cov_gr}
\alias{sample_cov_gr}
\title{sample_cov_gr}
\usage{
sample_cov_gr(psites, anno, ritpms)
}
\arguments{
\item{psites}{A GRanges object containing psites}

\item{anno}{An annotation object with a gene-transcript table}

\item{ritpms}{a vector of ribosoome densities}
}
\value{
a vector of normalized footprint densities
}
\description{
This function takes in a coverage GR and takes one out of each multimap
weighting according to the ritpms
}
