% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_div.R
\name{select_offsets}
\alias{select_offsets}
\title{Get a dataframe with p-site offsets from a datafram of KL divergence}
\usage{
select_offsets(kl_df, method = "a_max")
}
\arguments{
\item{kl_df}{dataframe with KL divergence per readlength, position in window
around codon}

\item{method}{means by which to select best offfset currently only possible value
is a_max - assumes the maximum KL divergence occurs under the A site}
}
\value{
a dataframe with p-site offsets per readlength, sample
}
\description{
Get a dataframe with p-site offsets from a datafram of KL divergence
}
\examples{
data(chr22_anno)
data(rpfs)
data(offsets_df)
covgrs <- list(sample1 = rpfs)
# note this doesn't work that well on a small subset
kl_df <- get_kl_df(covgrs, chr22_anno) #'
kl_offsets <- select_offsets(kl_df)
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
