% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{sort_grl_st}
\alias{sort_grl_st}
\title{Sort a GRanges list object with the sub-elements ordered 5' to 3'}
\usage{
sort_grl_st(grl)
}
\arguments{
\item{grl}{GRangesList; a GRangesList object}
}
\value{
A GRangeList object with the sub-elements ordered 5' to 3'
}
\description{
Sort a GRanges list object with the sub-elements ordered 5' to 3'
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
