% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{subset_annotation}
\alias{subset_annotation}
\title{Subset an annotation object with a vector of ORF ids}
\usage{
subset_annotation(anno, orfs)
}
\arguments{
\item{anno}{an annotation object}

\item{orfs}{A vector of ORF ids}
}
\value{
an annotation object, subsetted
}
\description{
Subset an annotation object with a vector of ORF ids
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
