% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{width1grs}
\alias{width1grs}
\title{given a granges object, convert it to width1 granges, preserving mcols}
\usage{
width1grs(gr)
}
\arguments{
\item{gr}{GRanges object, sorted}
}
\value{
A granges object with the each position in as it's own range
}
\description{
Given rpfs, and named vectors of starts and stops, this adds the 'phase'
to the reads
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
