% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_div.R
\name{get_metacodon_profs}
\alias{get_metacodon_profs}
\title{Get A Granges object with locations of each codon in the given cds}
\usage{
get_metacodon_profs(covgrs, anno, n_wind_l_ext = 45)
}
\arguments{
\item{covgrs}{A list of SimpleRleLists - RPF coverage}

\item{anno}{An annotation object}

\item{n_wind_l_ext}{The 5' extent of the window around the codon to return
Defaults to: 45}
}
\value{
A data frame with expected and actual rust score per position in
window, codon, readlength
}
\description{
Get A Granges object with locations of each codon in the given cds
}
\details{
gets profiles of read 5' end occurence around codons
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
