% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_TE.R
\name{get_psitecov}
\alias{get_psitecov}
\title{Generate a fake dataset of p-sites}
\usage{
get_psitecov(rpfs, offsets_df, anno)
}
\arguments{
\item{rpfs}{\itemize{
\item granges object with positions of footprints, in transcript
space
}}

\item{offsets_df}{\itemize{
\item a data frame with numeric columns readlen,phase,offset
}}

\item{anno}{\itemize{
\item an annotation object with at least one ORF per transcript
}}

\item{orfnames}{\itemize{
\item the ORFs to get sequence data for
}}
}
\value{
returns a data frame with columns orf_id, pos, codon
get a coverage RleList of psite coverage over a set of ORFs

a numeric SimpleRleList for each ORF with p-site density
}
\description{
This function useful for testing, generates a fake set of psites
given an annotation,
object
to get estimates of per-codon occupancy (which should be proportional to
dwell times)
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}

Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
