% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{get_trspace_cds}
\alias{get_trspace_cds}
\title{Check if a granges list of CDS have start codons}
\usage{
get_trspace_cds(cdsgrl, exonsgrl)
}
\arguments{
\item{exonsgrl}{exonsgrl; exons making up the space element is to be mapped
from.}

\item{trspacegr}{GRanges; an object in transcript space, to be mapped back
to the genome}
}
\value{
a granges object containing the coding sequence range for each
transcript
}
\description{
Check if a granges list of CDS have start codons
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
