% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{resize_grl}
\alias{resize_grl}
\title{Resize a GRangesList object, respecting exon boundaries when shortening}
\usage{
resize_grl(grl, gwidth, fix = "start", check = TRUE)
}
\arguments{
\item{grl}{GRangesList; a GRangesList object}

\item{gwidth}{GRangesList; integer/IntegerList to set as new width.}

\item{fix}{Position to fix to when resizing (start, stop, or center)}

\item{check}{Boolean; check if resulting GRanges extend beyond bounds}
}
\value{
A GRangesList object shortend/lengthened, respecting exon boundaries
}
\description{
Resize a GRangesList object, respecting exon boundaries when shortening
}
\author{
Dermot Harnett, \email{dermot.p.harnett@gmail.com}
}
\keyword{Ribostan}
